Higher Order Funcitons - Exercise
Create an extension on List using a higher order function that returns all
the numbers in the list that are divisible by 3. Start by creating an extension
function on List that takes an lambda on Int and applies it to each item in the
list. When the lambda returns zero, include the item in the output.